requirements.txt
Flask
opencv-python
mediapipe
numpy
Pillow